requirements.txt
altair
attrs 
blinker 
cachetools
certifi
charset-normalizer
click
entrypoints
gitdb
GitPython
greenlet
idna
Jinja2
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
numpy
packaging
pandas
pillow
plotly
protobuf
psycopg2-binary
pyarrow
pydeck
Pygments
python-dateutil
python-decouple
pytz
referencing
requests
rich
rpds-py
six
smmap
SQLAlchemy
streamlit
streamlit-aggrid
tenacity
toml
toolz
tornado
typing_extensions
tzdata
urllib3
watchdog
